Struct rusoto_glue::CreateDevEndpointResponse

pub struct CreateDevEndpointResponse {
    pub availability_zone: Option<String>,
    pub created_timestamp: Option<f64>,
    pub endpoint_name: Option<String>,
    pub extra_jars_s3_path: Option<String>,
    pub extra_python_libs_s3_path: Option<String>,
    pub failure_reason: Option<String>,
    pub number_of_nodes: Option<i64>,
    pub role_arn: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub status: Option<String>,
    pub subnet_id: Option<String>,
    pub vpc_id: Option<String>,
    pub yarn_endpoint_address: Option<String>,
}

Fields
availability_zone: Option<String>

The AWS availability zone where this DevEndpoint is located.

created_timestamp: Option<f64>

The point in time at which this DevEndpoint was created.

endpoint_name: Option<String>

The name assigned to the new DevEndpoint.

extra_jars_s3_path: Option<String>

Path to one or more Java Jars in an S3 bucket that will be loaded in your DevEndpoint.

extra_python_libs_s3_path: Option<String>

Path to one or more Python libraries in an S3 bucket that will be loaded in your DevEndpoint.

failure_reason: Option<String>

The reason for a current failure in this DevEndpoint.

number_of_nodes: Option<i64>

The number of nodes in this DevEndpoint.

role_arn: Option<String>

The AWS ARN of the role assigned to the new DevEndpoint.

security_group_ids: Option<Vec<String>>

The security groups assigned to the new DevEndpoint.

status: Option<String>

The current status of the new DevEndpoint.

subnet_id: Option<String>

The subnet ID assigned to the new DevEndpoint.

vpc_id: Option<String>

The ID of the VPC used by this DevEndpoint.

yarn_endpoint_address: Option<String>

The address of the YARN endpoint used by this DevEndpoint.