Struct rusoto_glue::CreateJobRequest

pub struct CreateJobRequest {
    pub allocated_capacity: Option<i64>,
    pub command: JobCommand,
    pub connections: Option<ConnectionsList>,
    pub default_arguments: Option<HashMap<String, String>>,
    pub description: Option<String>,
    pub execution_property: Option<ExecutionProperty>,
    pub log_uri: Option<String>,
    pub max_retries: Option<i64>,
    pub name: String,
    pub role: String,
}

Fields
allocated_capacity: Option<i64>

The number of capacity units allocated to this job.

command: JobCommand

The JobCommand that executes this job.

connections: Option<ConnectionsList>

The connections used for this job.

default_arguments: Option<HashMap<String, String>>

The default parameters for this job.

description: Option<String>

Description of the job.

execution_property: Option<ExecutionProperty>

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

log_uri: Option<String>

Location of the logs for this job.

max_retries: Option<i64>

The maximum number of times to retry this job if it fails.

name: String

The name you assign to this job.

role: String

The role associated with this job.