Struct rusoto_glue::DevEndpoint

pub struct DevEndpoint {
    pub arguments: Option<HashMap<String, String>>,
    pub availability_zone: Option<String>,
    pub created_timestamp: Option<f64>,
    pub endpoint_name: Option<String>,
    pub extra_jars_s3_path: Option<String>,
    pub extra_python_libs_s3_path: Option<String>,
    pub failure_reason: Option<String>,
    pub last_modified_timestamp: Option<f64>,
    pub last_update_status: Option<String>,
    pub number_of_nodes: Option<i64>,
    pub private_address: Option<String>,
    pub public_address: Option<String>,
    pub public_key: Option<String>,
    pub public_keys: Option<Vec<String>>,
    pub role_arn: Option<String>,
    pub security_configuration: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub status: Option<String>,
    pub subnet_id: Option<String>,
    pub vpc_id: Option<String>,
    pub yarn_endpoint_address: Option<String>,
    pub zeppelin_remote_spark_interpreter_port: Option<i64>,
}

A development endpoint where a developer can remotely debug ETL scripts.

Fields
arguments: Option<HashMap<String, String>>

A map of arguments used to configure the DevEndpoint.

Note that currently, we only support "--enable-glue-datacatalog": "" as a valid argument.

availability_zone: Option<String>

The AWS availability zone where this DevEndpoint is located.

created_timestamp: Option<f64>

The point in time at which this DevEndpoint was created.

endpoint_name: Option<String>

The name of the DevEndpoint.

extra_jars_s3_path: Option<String>

Path to one or more Java Jars in an S3 bucket that should be loaded in your DevEndpoint.

Please note that only pure Java/Scala libraries can currently be used on a DevEndpoint.

extra_python_libs_s3_path: Option<String>

Path(s) to one or more Python libraries in an S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

Please note that only pure Python libraries can currently be used on a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not yet supported.

failure_reason: Option<String>

The reason for a current failure in this DevEndpoint.

last_modified_timestamp: Option<f64>

The point in time at which this DevEndpoint was last modified.

last_update_status: Option<String>

The status of the last update.

number_of_nodes: Option<i64>

The number of AWS Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

private_address: Option<String>

A private IP address to access the DevEndpoint within a VPC, if the DevEndpoint is created within one. The PrivateAddress field is present only when you create the DevEndpoint within your virtual private cloud (VPC).

public_address: Option<String>

The public IP address used by this DevEndpoint. The PublicAddress field is present only when you create a non-VPC (virtual private cloud) DevEndpoint.

public_key: Option<String>

The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility, as the recommended attribute to use is public keys.

public_keys: Option<Vec<String>>

A list of public keys to be used by the DevEndpoints for authentication. The use of this attribute is preferred over a single public key because the public keys allow you to have a different private key per client.

If you previously created an endpoint with a public key, you must remove that key to be able to set a list of public keys: call the UpdateDevEndpoint API with the public key content in the deletePublicKeys attribute, and the list of new keys in the addPublicKeys attribute.

role_arn: Option<String>

The AWS ARN of the IAM role used in this DevEndpoint.

security_configuration: Option<String>

The name of the SecurityConfiguration structure to be used with this DevEndpoint.

security_group_ids: Option<Vec<String>>

A list of security group identifiers used in this DevEndpoint.

status: Option<String>

The current status of this DevEndpoint.

subnet_id: Option<String>

The subnet ID for this DevEndpoint.

vpc_id: Option<String>

The ID of the virtual private cloud (VPC) used by this DevEndpoint.

yarn_endpoint_address: Option<String>

The YARN endpoint address used by this DevEndpoint.

zeppelin_remote_spark_interpreter_port: Option<i64>

The Apache Zeppelin port for the remote Apache Spark interpreter.