Struct rusoto_glue::GlueClient

pub struct GlueClient { /* fields omitted */ }

A client for the AWS Glue API.

Methods

impl GlueClient

pub fn new(region: Region) -> GlueClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> GlueClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,