Struct rusoto_glue::JobRun

pub struct JobRun {
    pub allocated_capacity: Option<i64>,
    pub arguments: Option<HashMap<String, String>>,
    pub attempt: Option<i64>,
    pub completed_on: Option<f64>,
    pub error_message: Option<String>,
    pub id: Option<String>,
    pub job_name: Option<String>,
    pub job_run_state: Option<String>,
    pub last_modified_on: Option<f64>,
    pub predecessor_runs: Option<Vec<Predecessor>>,
    pub previous_run_id: Option<String>,
    pub started_on: Option<f64>,
    pub trigger_name: Option<String>,
}

Contains information about a job run.

Fields
allocated_capacity: Option<i64>

The amount of infrastructure capacity allocated to this job run.

arguments: Option<HashMap<String, String>>

The job arguments associated with this run.

attempt: Option<i64>

The number or the attempt to run this job.

completed_on: Option<f64>

The date and time this job run completed.

error_message: Option<String>

An error message associated with this job run.

id: Option<String>

The ID of this job run.

job_name: Option<String>

The name of the job being run.

job_run_state: Option<String>

The current state of the job run.

last_modified_on: Option<f64>

The last time this job run was modified.

predecessor_runs: Option<Vec<Predecessor>>

A list of predecessors to this job run.

previous_run_id: Option<String>

The ID of the previous run of this job.

started_on: Option<f64>

The date and time at which this job run was started.

trigger_name: Option<String>

The name of the trigger for this job run.