Struct rusoto_glue::Partition

pub struct Partition {
    pub creation_time: Option<f64>,
    pub database_name: Option<String>,
    pub last_access_time: Option<f64>,
    pub last_analyzed_time: Option<f64>,
    pub parameters: Option<HashMap<String, String>>,
    pub storage_descriptor: Option<StorageDescriptor>,
    pub table_name: Option<String>,
    pub values: Option<Vec<String>>,
}

Represents a slice of table data.

Fields
creation_time: Option<f64>

The time at which the partition was created.

database_name: Option<String>

The name of the catalog database where the table in question is located.

last_access_time: Option<f64>

The last time at which the partition was accessed.

last_analyzed_time: Option<f64>

The last time at which column statistics were computed for this partition.

parameters: Option<HashMap<String, String>>

Partition parameters, in the form of a list of key-value pairs.

storage_descriptor: Option<StorageDescriptor>

Provides information about the physical location where the partition is stored.

table_name: Option<String>

The name of the table in question.

values: Option<Vec<String>>

The values of the partition.