Struct rusoto_glue::StorageDescriptor

pub struct StorageDescriptor {
    pub bucket_columns: Option<Vec<String>>,
    pub columns: Option<Vec<Column>>,
    pub compressed: Option<bool>,
    pub input_format: Option<String>,
    pub location: Option<String>,
    pub number_of_buckets: Option<i64>,
    pub output_format: Option<String>,
    pub parameters: Option<HashMap<String, String>>,
    pub serde_info: Option<SerDeInfo>,
    pub skewed_info: Option<SkewedInfo>,
    pub sort_columns: Option<Vec<Order>>,
    pub stored_as_sub_directories: Option<bool>,
}

Describes the physical storage of table data.

Fields
bucket_columns: Option<Vec<String>>

A list of reducer grouping columns, clustering columns, and bucketing columns in the table.

columns: Option<Vec<Column>>

A list of the Columns in the table.

compressed: Option<bool>

True if the data in the table is compressed, or False if not.

input_format: Option<String>

The input format: SequenceFileInputFormat (binary), or TextInputFormat, or a custom format.

location: Option<String>

The physical location of the table. By default this takes the form of the warehouse location, followed by the database location in the warehouse, followed by the table name.

number_of_buckets: Option<i64>

Must be specified if the table contains any dimension columns.

output_format: Option<String>

The output format: SequenceFileOutputFormat (binary), or IgnoreKeyTextOutputFormat, or a custom format.

parameters: Option<HashMap<String, String>>

User-supplied properties in key-value form.

serde_info: Option<SerDeInfo>

Serialization/deserialization (SerDe) information.

skewed_info: Option<SkewedInfo>

Information about values that appear very frequently in a column (skewed values).

sort_columns: Option<Vec<Order>>

A list specifying the sort order of each bucket in the table.

stored_as_sub_directories: Option<bool>

True if the table data is stored in subdirectories, or False if not.