Trait rusoto_glue::Glue

pub trait Glue {
    fn batch_create_partition(
        &self, 
        input: BatchCreatePartitionRequest
    ) -> RusotoFuture<BatchCreatePartitionResponse, BatchCreatePartitionError>;

    fn batch_delete_connection(
        &self, 
        input: BatchDeleteConnectionRequest
    ) -> RusotoFuture<BatchDeleteConnectionResponse, BatchDeleteConnectionError>;

    fn batch_delete_partition(
        &self, 
        input: BatchDeletePartitionRequest
    ) -> RusotoFuture<BatchDeletePartitionResponse, BatchDeletePartitionError>;

    fn batch_delete_table(
        &self, 
        input: BatchDeleteTableRequest
    ) -> RusotoFuture<BatchDeleteTableResponse, BatchDeleteTableError>;

    fn batch_delete_table_version(
        &self, 
        input: BatchDeleteTableVersionRequest
    ) -> RusotoFuture<BatchDeleteTableVersionResponse, BatchDeleteTableVersionError>;

    fn batch_get_crawlers(
        &self, 
        input: BatchGetCrawlersRequest
    ) -> RusotoFuture<BatchGetCrawlersResponse, BatchGetCrawlersError>;

    fn batch_get_dev_endpoints(
        &self, 
        input: BatchGetDevEndpointsRequest
    ) -> RusotoFuture<BatchGetDevEndpointsResponse, BatchGetDevEndpointsError>;

    fn batch_get_jobs(
        &self, 
        input: BatchGetJobsRequest
    ) -> RusotoFuture<BatchGetJobsResponse, BatchGetJobsError>;

    fn batch_get_partition(
        &self, 
        input: BatchGetPartitionRequest
    ) -> RusotoFuture<BatchGetPartitionResponse, BatchGetPartitionError>;

    fn batch_get_triggers(
        &self, 
        input: BatchGetTriggersRequest
    ) -> RusotoFuture<BatchGetTriggersResponse, BatchGetTriggersError>;

    fn batch_stop_job_run(
        &self, 
        input: BatchStopJobRunRequest
    ) -> RusotoFuture<BatchStopJobRunResponse, GlueBatchStopJobRunError>;

    fn create_classifier(
        &self, 
        input: CreateClassifierRequest
    ) -> RusotoFuture<CreateClassifierResponse, CreateClassifierError>;

    fn create_connection(
        &self, 
        input: CreateConnectionRequest
    ) -> RusotoFuture<CreateConnectionResponse, CreateConnectionError>;

    fn create_crawler(
        &self, 
        input: CreateCrawlerRequest
    ) -> RusotoFuture<CreateCrawlerResponse, CreateCrawlerError>;

    fn create_database(
        &self, 
        input: CreateDatabaseRequest
    ) -> RusotoFuture<CreateDatabaseResponse, CreateDatabaseError>;

    fn create_dev_endpoint(
        &self, 
        input: CreateDevEndpointRequest
    ) -> RusotoFuture<CreateDevEndpointResponse, CreateDevEndpointError>;

    fn create_job(
        &self, 
        input: CreateJobRequest
    ) -> RusotoFuture<CreateJobResponse, CreateJobError>;

    fn create_partition(
        &self, 
        input: CreatePartitionRequest
    ) -> RusotoFuture<CreatePartitionResponse, CreatePartitionError>;

    fn create_script(
        &self, 
        input: CreateScriptRequest
    ) -> RusotoFuture<CreateScriptResponse, CreateScriptError>;

    fn create_security_configuration(
        &self, 
        input: CreateSecurityConfigurationRequest
    ) -> RusotoFuture<CreateSecurityConfigurationResponse, CreateSecurityConfigurationError>;

    fn create_table(
        &self, 
        input: CreateTableRequest
    ) -> RusotoFuture<CreateTableResponse, CreateTableError>;

    fn create_trigger(
        &self, 
        input: CreateTriggerRequest
    ) -> RusotoFuture<CreateTriggerResponse, CreateTriggerError>;

    fn create_user_defined_function(
        &self, 
        input: CreateUserDefinedFunctionRequest
    ) -> RusotoFuture<CreateUserDefinedFunctionResponse, CreateUserDefinedFunctionError>;

    fn delete_classifier(
        &self, 
        input: DeleteClassifierRequest
    ) -> RusotoFuture<DeleteClassifierResponse, DeleteClassifierError>;

    fn delete_connection(
        &self, 
        input: DeleteConnectionRequest
    ) -> RusotoFuture<DeleteConnectionResponse, DeleteConnectionError>;

    fn delete_crawler(
        &self, 
        input: DeleteCrawlerRequest
    ) -> RusotoFuture<DeleteCrawlerResponse, DeleteCrawlerError>;

    fn delete_database(
        &self, 
        input: DeleteDatabaseRequest
    ) -> RusotoFuture<DeleteDatabaseResponse, DeleteDatabaseError>;

    fn delete_dev_endpoint(
        &self, 
        input: DeleteDevEndpointRequest
    ) -> RusotoFuture<DeleteDevEndpointResponse, DeleteDevEndpointError>;

    fn delete_job(
        &self, 
        input: DeleteJobRequest
    ) -> RusotoFuture<DeleteJobResponse, DeleteJobError>;

    fn delete_partition(
        &self, 
        input: DeletePartitionRequest
    ) -> RusotoFuture<DeletePartitionResponse, DeletePartitionError>;

    fn delete_resource_policy(
        &self, 
        input: DeleteResourcePolicyRequest
    ) -> RusotoFuture<DeleteResourcePolicyResponse, DeleteResourcePolicyError>;

    fn delete_security_configuration(
        &self, 
        input: DeleteSecurityConfigurationRequest
    ) -> RusotoFuture<DeleteSecurityConfigurationResponse, DeleteSecurityConfigurationError>;

    fn delete_table(
        &self, 
        input: DeleteTableRequest
    ) -> RusotoFuture<DeleteTableResponse, DeleteTableError>;

    fn delete_table_version(
        &self, 
        input: DeleteTableVersionRequest
    ) -> RusotoFuture<DeleteTableVersionResponse, DeleteTableVersionError>;

    fn delete_trigger(
        &self, 
        input: DeleteTriggerRequest
    ) -> RusotoFuture<DeleteTriggerResponse, DeleteTriggerError>;

    fn delete_user_defined_function(
        &self, 
        input: DeleteUserDefinedFunctionRequest
    ) -> RusotoFuture<DeleteUserDefinedFunctionResponse, DeleteUserDefinedFunctionError>;

    fn get_catalog_import_status(
        &self, 
        input: GetCatalogImportStatusRequest
    ) -> RusotoFuture<GetCatalogImportStatusResponse, GetCatalogImportStatusError>;

    fn get_classifier(
        &self, 
        input: GetClassifierRequest
    ) -> RusotoFuture<GetClassifierResponse, GetClassifierError>;

    fn get_classifiers(
        &self, 
        input: GetClassifiersRequest
    ) -> RusotoFuture<GetClassifiersResponse, GetClassifiersError>;

    fn get_connection(
        &self, 
        input: GetConnectionRequest
    ) -> RusotoFuture<GetConnectionResponse, GetConnectionError>;

    fn get_connections(
        &self, 
        input: GetConnectionsRequest
    ) -> RusotoFuture<GetConnectionsResponse, GetConnectionsError>;

    fn get_crawler(
        &self, 
        input: GetCrawlerRequest
    ) -> RusotoFuture<GetCrawlerResponse, GetCrawlerError>;

    fn get_crawler_metrics(
        &self, 
        input: GetCrawlerMetricsRequest
    ) -> RusotoFuture<GetCrawlerMetricsResponse, GetCrawlerMetricsError>;

    fn get_crawlers(
        &self, 
        input: GetCrawlersRequest
    ) -> RusotoFuture<GetCrawlersResponse, GetCrawlersError>;

    fn get_data_catalog_encryption_settings(
        &self, 
        input: GetDataCatalogEncryptionSettingsRequest
    ) -> RusotoFuture<GetDataCatalogEncryptionSettingsResponse, GetDataCatalogEncryptionSettingsError>;

    fn get_database(
        &self, 
        input: GetDatabaseRequest
    ) -> RusotoFuture<GetDatabaseResponse, GetDatabaseError>;

    fn get_databases(
        &self, 
        input: GetDatabasesRequest
    ) -> RusotoFuture<GetDatabasesResponse, GetDatabasesError>;

    fn get_dataflow_graph(
        &self, 
        input: GetDataflowGraphRequest
    ) -> RusotoFuture<GetDataflowGraphResponse, GetDataflowGraphError>;

    fn get_dev_endpoint(
        &self, 
        input: GetDevEndpointRequest
    ) -> RusotoFuture<GetDevEndpointResponse, GetDevEndpointError>;

    fn get_dev_endpoints(
        &self, 
        input: GetDevEndpointsRequest
    ) -> RusotoFuture<GetDevEndpointsResponse, GetDevEndpointsError>;

    fn get_job(
        &self, 
        input: GetJobRequest
    ) -> RusotoFuture<GetJobResponse, GetJobError>;

    fn get_job_run(
        &self, 
        input: GetJobRunRequest
    ) -> RusotoFuture<GetJobRunResponse, GetJobRunError>;

    fn get_job_runs(
        &self, 
        input: GetJobRunsRequest
    ) -> RusotoFuture<GetJobRunsResponse, GetJobRunsError>;

    fn get_jobs(
        &self, 
        input: GetJobsRequest
    ) -> RusotoFuture<GetJobsResponse, GetJobsError>;

    fn get_mapping(
        &self, 
        input: GetMappingRequest
    ) -> RusotoFuture<GetMappingResponse, GetMappingError>;

    fn get_partition(
        &self, 
        input: GetPartitionRequest
    ) -> RusotoFuture<GetPartitionResponse, GetPartitionError>;

    fn get_partitions(
        &self, 
        input: GetPartitionsRequest
    ) -> RusotoFuture<GetPartitionsResponse, GetPartitionsError>;

    fn get_plan(
        &self, 
        input: GetPlanRequest
    ) -> RusotoFuture<GetPlanResponse, GetPlanError>;

    fn get_resource_policy(
        &self
    ) -> RusotoFuture<GetResourcePolicyResponse, GetResourcePolicyError>;

    fn get_security_configuration(
        &self, 
        input: GetSecurityConfigurationRequest
    ) -> RusotoFuture<GetSecurityConfigurationResponse, GetSecurityConfigurationError>;

    fn get_security_configurations(
        &self, 
        input: GetSecurityConfigurationsRequest
    ) -> RusotoFuture<GetSecurityConfigurationsResponse, GetSecurityConfigurationsError>;

    fn get_table(
        &self, 
        input: GetTableRequest
    ) -> RusotoFuture<GetTableResponse, GetTableError>;

    fn get_table_version(
        &self, 
        input: GetTableVersionRequest
    ) -> RusotoFuture<GetTableVersionResponse, GetTableVersionError>;

    fn get_table_versions(
        &self, 
        input: GetTableVersionsRequest
    ) -> RusotoFuture<GetTableVersionsResponse, GetTableVersionsError>;

    fn get_tables(
        &self, 
        input: GetTablesRequest
    ) -> RusotoFuture<GetTablesResponse, GetTablesError>;

    fn get_tags(
        &self, 
        input: GetTagsRequest
    ) -> RusotoFuture<GetTagsResponse, GetTagsError>;

    fn get_trigger(
        &self, 
        input: GetTriggerRequest
    ) -> RusotoFuture<GetTriggerResponse, GetTriggerError>;

    fn get_triggers(
        &self, 
        input: GetTriggersRequest
    ) -> RusotoFuture<GetTriggersResponse, GetTriggersError>;

    fn get_user_defined_function(
        &self, 
        input: GetUserDefinedFunctionRequest
    ) -> RusotoFuture<GetUserDefinedFunctionResponse, GetUserDefinedFunctionError>;

    fn get_user_defined_functions(
        &self, 
        input: GetUserDefinedFunctionsRequest
    ) -> RusotoFuture<GetUserDefinedFunctionsResponse, GetUserDefinedFunctionsError>;

    fn import_catalog_to_glue(
        &self, 
        input: ImportCatalogToGlueRequest
    ) -> RusotoFuture<ImportCatalogToGlueResponse, ImportCatalogToGlueError>;

    fn list_crawlers(
        &self, 
        input: ListCrawlersRequest
    ) -> RusotoFuture<ListCrawlersResponse, ListCrawlersError>;

    fn list_dev_endpoints(
        &self, 
        input: ListDevEndpointsRequest
    ) -> RusotoFuture<ListDevEndpointsResponse, ListDevEndpointsError>;

    fn list_jobs(
        &self, 
        input: ListJobsRequest
    ) -> RusotoFuture<ListJobsResponse, ListJobsError>;

    fn list_triggers(
        &self, 
        input: ListTriggersRequest
    ) -> RusotoFuture<ListTriggersResponse, ListTriggersError>;

    fn put_data_catalog_encryption_settings(
        &self, 
        input: PutDataCatalogEncryptionSettingsRequest
    ) -> RusotoFuture<PutDataCatalogEncryptionSettingsResponse, PutDataCatalogEncryptionSettingsError>;

    fn put_resource_policy(
        &self, 
        input: PutResourcePolicyRequest
    ) -> RusotoFuture<PutResourcePolicyResponse, PutResourcePolicyError>;

    fn reset_job_bookmark(
        &self, 
        input: ResetJobBookmarkRequest
    ) -> RusotoFuture<ResetJobBookmarkResponse, ResetJobBookmarkError>;

    fn start_crawler(
        &self, 
        input: StartCrawlerRequest
    ) -> RusotoFuture<StartCrawlerResponse, StartCrawlerError>;

    fn start_crawler_schedule(
        &self, 
        input: StartCrawlerScheduleRequest
    ) -> RusotoFuture<StartCrawlerScheduleResponse, StartCrawlerScheduleError>;

    fn start_job_run(
        &self, 
        input: StartJobRunRequest
    ) -> RusotoFuture<StartJobRunResponse, StartJobRunError>;

    fn start_trigger(
        &self, 
        input: StartTriggerRequest
    ) -> RusotoFuture<StartTriggerResponse, StartTriggerError>;

    fn stop_crawler(
        &self, 
        input: StopCrawlerRequest
    ) -> RusotoFuture<StopCrawlerResponse, StopCrawlerError>;

    fn stop_crawler_schedule(
        &self, 
        input: StopCrawlerScheduleRequest
    ) -> RusotoFuture<StopCrawlerScheduleResponse, StopCrawlerScheduleError>;

    fn stop_trigger(
        &self, 
        input: StopTriggerRequest
    ) -> RusotoFuture<StopTriggerResponse, StopTriggerError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_classifier(
        &self, 
        input: UpdateClassifierRequest
    ) -> RusotoFuture<UpdateClassifierResponse, UpdateClassifierError>;

    fn update_connection(
        &self, 
        input: UpdateConnectionRequest
    ) -> RusotoFuture<UpdateConnectionResponse, UpdateConnectionError>;

    fn update_crawler(
        &self, 
        input: UpdateCrawlerRequest
    ) -> RusotoFuture<UpdateCrawlerResponse, UpdateCrawlerError>;

    fn update_crawler_schedule(
        &self, 
        input: UpdateCrawlerScheduleRequest
    ) -> RusotoFuture<UpdateCrawlerScheduleResponse, UpdateCrawlerScheduleError>;

    fn update_database(
        &self, 
        input: UpdateDatabaseRequest
    ) -> RusotoFuture<UpdateDatabaseResponse, UpdateDatabaseError>;

    fn update_dev_endpoint(
        &self, 
        input: UpdateDevEndpointRequest
    ) -> RusotoFuture<UpdateDevEndpointResponse, UpdateDevEndpointError>;

    fn update_job(
        &self, 
        input: UpdateJobRequest
    ) -> RusotoFuture<UpdateJobResponse, UpdateJobError>;

    fn update_partition(
        &self, 
        input: UpdatePartitionRequest
    ) -> RusotoFuture<UpdatePartitionResponse, UpdatePartitionError>;

    fn update_table(
        &self, 
        input: UpdateTableRequest
    ) -> RusotoFuture<UpdateTableResponse, UpdateTableError>;

    fn update_trigger(
        &self, 
        input: UpdateTriggerRequest
    ) -> RusotoFuture<UpdateTriggerResponse, UpdateTriggerError>;

    fn update_user_defined_function(
        &self, 
        input: UpdateUserDefinedFunctionRequest
    ) -> RusotoFuture<UpdateUserDefinedFunctionResponse, UpdateUserDefinedFunctionError>;
}

Trait representing the capabilities of the AWS Glue API. AWS Glue clients implement this trait.

Required methods

fn batch_create_partition(
&self,
input: BatchCreatePartitionRequest
) -> RusotoFuture<BatchCreatePartitionResponse, BatchCreatePartitionError>

Creates one or more partitions in a batch operation.

fn batch_delete_connection(
&self,
input: BatchDeleteConnectionRequest
) -> RusotoFuture<BatchDeleteConnectionResponse, BatchDeleteConnectionError>

Deletes a list of connection definitions from the Data Catalog.

fn batch_delete_partition(
&self,
input: BatchDeletePartitionRequest
) -> RusotoFuture<BatchDeletePartitionResponse, BatchDeletePartitionError>

Deletes one or more partitions in a batch operation.

fn batch_delete_table(
&self,
input: BatchDeleteTableRequest
) -> RusotoFuture<BatchDeleteTableResponse, BatchDeleteTableError>

Deletes multiple tables at once.

After completing this operation, you will no longer have access to the table versions and partitions that belong to the deleted table. AWS Glue deletes these "orphaned" resources asynchronously in a timely manner, at the discretion of the service.

To ensure immediate deletion of all related resources, before calling BatchDeleteTable, use DeleteTableVersion or BatchDeleteTableVersion, and DeletePartition or BatchDeletePartition, to delete any resources that belong to the table.

fn batch_delete_table_version(
&self,
input: BatchDeleteTableVersionRequest
) -> RusotoFuture<BatchDeleteTableVersionResponse, BatchDeleteTableVersionError>

Deletes a specified batch of versions of a table.

fn batch_get_crawlers(
&self,
input: BatchGetCrawlersRequest
) -> RusotoFuture<BatchGetCrawlersResponse, BatchGetCrawlersError>

Returns a list of resource metadata for a given list of crawler names. After calling the ListCrawlers operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

fn batch_get_dev_endpoints(
&self,
input: BatchGetDevEndpointsRequest
) -> RusotoFuture<BatchGetDevEndpointsResponse, BatchGetDevEndpointsError>

Returns a list of resource metadata for a given list of DevEndpoint names. After calling the ListDevEndpoints operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

fn batch_get_jobs(
&self,
input: BatchGetJobsRequest
) -> RusotoFuture<BatchGetJobsResponse, BatchGetJobsError>

Returns a list of resource metadata for a given list of job names. After calling the ListJobs operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

fn batch_get_partition(
&self,
input: BatchGetPartitionRequest
) -> RusotoFuture<BatchGetPartitionResponse, BatchGetPartitionError>

Retrieves partitions in a batch request.

fn batch_get_triggers(
&self,
input: BatchGetTriggersRequest
) -> RusotoFuture<BatchGetTriggersResponse, BatchGetTriggersError>

Returns a list of resource metadata for a given list of trigger names. After calling the ListTriggers operation, you can call this operation to access the data to which you have been granted permissions. This operation supports all IAM permissions, including permission conditions that uses tags.

fn batch_stop_job_run(
&self,
input: BatchStopJobRunRequest
) -> RusotoFuture<BatchStopJobRunResponse, GlueBatchStopJobRunError>

Stops one or more job runs for a specified job definition.

fn create_classifier(
&self,
input: CreateClassifierRequest
) -> RusotoFuture<CreateClassifierResponse, CreateClassifierError>

Creates a classifier in the user's account. This can be a GrokClassifier, an XMLClassifier, a JsonClassifier, or a CsvClassifier, depending on which field of the request is present.

fn create_connection(
&self,
input: CreateConnectionRequest
) -> RusotoFuture<CreateConnectionResponse, CreateConnectionError>

Creates a connection definition in the Data Catalog.

fn create_crawler(
&self,
input: CreateCrawlerRequest
) -> RusotoFuture<CreateCrawlerResponse, CreateCrawlerError>

Creates a new crawler with specified targets, role, configuration, and optional schedule. At least one crawl target must be specified, in the s3Targets field, the jdbcTargets field, or the DynamoDBTargets field.

fn create_database(
&self,
input: CreateDatabaseRequest
) -> RusotoFuture<CreateDatabaseResponse, CreateDatabaseError>

Creates a new database in a Data Catalog.

fn create_dev_endpoint(
&self,
input: CreateDevEndpointRequest
) -> RusotoFuture<CreateDevEndpointResponse, CreateDevEndpointError>

Creates a new DevEndpoint.

fn create_job(
&self,
input: CreateJobRequest
) -> RusotoFuture<CreateJobResponse, CreateJobError>

Creates a new job definition.

fn create_partition(
&self,
input: CreatePartitionRequest
) -> RusotoFuture<CreatePartitionResponse, CreatePartitionError>

Creates a new partition.

fn create_script(
&self,
input: CreateScriptRequest
) -> RusotoFuture<CreateScriptResponse, CreateScriptError>

Transforms a directed acyclic graph (DAG) into code.

fn create_security_configuration(
&self,
input: CreateSecurityConfigurationRequest
) -> RusotoFuture<CreateSecurityConfigurationResponse, CreateSecurityConfigurationError>

Creates a new security configuration.

fn create_table(
&self,
input: CreateTableRequest
) -> RusotoFuture<CreateTableResponse, CreateTableError>

Creates a new table definition in the Data Catalog.

fn create_trigger(
&self,
input: CreateTriggerRequest
) -> RusotoFuture<CreateTriggerResponse, CreateTriggerError>

Creates a new trigger.

fn create_user_defined_function(
&self,
input: CreateUserDefinedFunctionRequest
) -> RusotoFuture<CreateUserDefinedFunctionResponse, CreateUserDefinedFunctionError>

Creates a new function definition in the Data Catalog.

fn delete_classifier(
&self,
input: DeleteClassifierRequest
) -> RusotoFuture<DeleteClassifierResponse, DeleteClassifierError>

Removes a classifier from the Data Catalog.

fn delete_connection(
&self,
input: DeleteConnectionRequest
) -> RusotoFuture<DeleteConnectionResponse, DeleteConnectionError>

Deletes a connection from the Data Catalog.

fn delete_crawler(
&self,
input: DeleteCrawlerRequest
) -> RusotoFuture<DeleteCrawlerResponse, DeleteCrawlerError>

Removes a specified crawler from the AWS Glue Data Catalog, unless the crawler state is RUNNING.

fn delete_database(
&self,
input: DeleteDatabaseRequest
) -> RusotoFuture<DeleteDatabaseResponse, DeleteDatabaseError>

Removes a specified Database from a Data Catalog.

After completing this operation, you will no longer have access to the tables (and all table versions and partitions that might belong to the tables) and the user-defined functions in the deleted database. AWS Glue deletes these "orphaned" resources asynchronously in a timely manner, at the discretion of the service.

To ensure immediate deletion of all related resources, before calling DeleteDatabase, use DeleteTableVersion or BatchDeleteTableVersion, DeletePartition or BatchDeletePartition, DeleteUserDefinedFunction, and DeleteTable or BatchDeleteTable, to delete any resources that belong to the database.

fn delete_dev_endpoint(
&self,
input: DeleteDevEndpointRequest
) -> RusotoFuture<DeleteDevEndpointResponse, DeleteDevEndpointError>

Deletes a specified DevEndpoint.

fn delete_job(
&self,
input: DeleteJobRequest
) -> RusotoFuture<DeleteJobResponse, DeleteJobError>

Deletes a specified job definition. If the job definition is not found, no exception is thrown.

fn delete_partition(
&self,
input: DeletePartitionRequest
) -> RusotoFuture<DeletePartitionResponse, DeletePartitionError>

Deletes a specified partition.

fn delete_resource_policy(
&self,
input: DeleteResourcePolicyRequest
) -> RusotoFuture<DeleteResourcePolicyResponse, DeleteResourcePolicyError>

Deletes a specified policy.

fn delete_security_configuration(
&self,
input: DeleteSecurityConfigurationRequest
) -> RusotoFuture<DeleteSecurityConfigurationResponse, DeleteSecurityConfigurationError>

Deletes a specified security configuration.

fn delete_table(
&self,
input: DeleteTableRequest
) -> RusotoFuture<DeleteTableResponse, DeleteTableError>

Removes a table definition from the Data Catalog.

After completing this operation, you will no longer have access to the table versions and partitions that belong to the deleted table. AWS Glue deletes these "orphaned" resources asynchronously in a timely manner, at the discretion of the service.

To ensure immediate deletion of all related resources, before calling DeleteTable, use DeleteTableVersion or BatchDeleteTableVersion, and DeletePartition or BatchDeletePartition, to delete any resources that belong to the table.

fn delete_table_version(
&self,
input: DeleteTableVersionRequest
) -> RusotoFuture<DeleteTableVersionResponse, DeleteTableVersionError>

Deletes a specified version of a table.

fn delete_trigger(
&self,
input: DeleteTriggerRequest
) -> RusotoFuture<DeleteTriggerResponse, DeleteTriggerError>

Deletes a specified trigger. If the trigger is not found, no exception is thrown.

fn delete_user_defined_function(
&self,
input: DeleteUserDefinedFunctionRequest
) -> RusotoFuture<DeleteUserDefinedFunctionResponse, DeleteUserDefinedFunctionError>

Deletes an existing function definition from the Data Catalog.

fn get_catalog_import_status(
&self,
input: GetCatalogImportStatusRequest
) -> RusotoFuture<GetCatalogImportStatusResponse, GetCatalogImportStatusError>

Retrieves the status of a migration operation.

fn get_classifier(
&self,
input: GetClassifierRequest
) -> RusotoFuture<GetClassifierResponse, GetClassifierError>

Retrieve a classifier by name.

fn get_classifiers(
&self,
input: GetClassifiersRequest
) -> RusotoFuture<GetClassifiersResponse, GetClassifiersError>

Lists all classifier objects in the Data Catalog.

fn get_connection(
&self,
input: GetConnectionRequest
) -> RusotoFuture<GetConnectionResponse, GetConnectionError>

Retrieves a connection definition from the Data Catalog.

fn get_connections(
&self,
input: GetConnectionsRequest
) -> RusotoFuture<GetConnectionsResponse, GetConnectionsError>

Retrieves a list of connection definitions from the Data Catalog.

fn get_crawler(
&self,
input: GetCrawlerRequest
) -> RusotoFuture<GetCrawlerResponse, GetCrawlerError>

Retrieves metadata for a specified crawler.

fn get_crawler_metrics(
&self,
input: GetCrawlerMetricsRequest
) -> RusotoFuture<GetCrawlerMetricsResponse, GetCrawlerMetricsError>

Retrieves metrics about specified crawlers.

fn get_crawlers(
&self,
input: GetCrawlersRequest
) -> RusotoFuture<GetCrawlersResponse, GetCrawlersError>

Retrieves metadata for all crawlers defined in the customer account.

fn get_data_catalog_encryption_settings(
&self,
input: GetDataCatalogEncryptionSettingsRequest
) -> RusotoFuture<GetDataCatalogEncryptionSettingsResponse, GetDataCatalogEncryptionSettingsError>

Retrieves the security configuration for a specified catalog.

fn get_database(
&self,
input: GetDatabaseRequest
) -> RusotoFuture<GetDatabaseResponse, GetDatabaseError>

Retrieves the definition of a specified database.

fn get_databases(
&self,
input: GetDatabasesRequest
) -> RusotoFuture<GetDatabasesResponse, GetDatabasesError>

Retrieves all Databases defined in a given Data Catalog.

fn get_dataflow_graph(
&self,
input: GetDataflowGraphRequest
) -> RusotoFuture<GetDataflowGraphResponse, GetDataflowGraphError>

Transforms a Python script into a directed acyclic graph (DAG).

fn get_dev_endpoint(
&self,
input: GetDevEndpointRequest
) -> RusotoFuture<GetDevEndpointResponse, GetDevEndpointError>

Retrieves information about a specified DevEndpoint.

When you create a development endpoint in a virtual private cloud (VPC), AWS Glue returns only a private IP address, and the public IP address field is not populated. When you create a non-VPC development endpoint, AWS Glue returns only a public IP address.

fn get_dev_endpoints(
&self,
input: GetDevEndpointsRequest
) -> RusotoFuture<GetDevEndpointsResponse, GetDevEndpointsError>

Retrieves all the DevEndpoints in this AWS account.

When you create a development endpoint in a virtual private cloud (VPC), AWS Glue returns only a private IP address and the public IP address field is not populated. When you create a non-VPC development endpoint, AWS Glue returns only a public IP address.

fn get_job(
&self,
input: GetJobRequest
) -> RusotoFuture<GetJobResponse, GetJobError>

Retrieves an existing job definition.

fn get_job_run(
&self,
input: GetJobRunRequest
) -> RusotoFuture<GetJobRunResponse, GetJobRunError>

Retrieves the metadata for a given job run.

fn get_job_runs(
&self,
input: GetJobRunsRequest
) -> RusotoFuture<GetJobRunsResponse, GetJobRunsError>

Retrieves metadata for all runs of a given job definition.

fn get_jobs(
&self,
input: GetJobsRequest
) -> RusotoFuture<GetJobsResponse, GetJobsError>

Retrieves all current job definitions.

fn get_mapping(
&self,
input: GetMappingRequest
) -> RusotoFuture<GetMappingResponse, GetMappingError>

Creates mappings.

fn get_partition(
&self,
input: GetPartitionRequest
) -> RusotoFuture<GetPartitionResponse, GetPartitionError>

Retrieves information about a specified partition.

fn get_partitions(
&self,
input: GetPartitionsRequest
) -> RusotoFuture<GetPartitionsResponse, GetPartitionsError>

Retrieves information about the partitions in a table.

fn get_plan(
&self,
input: GetPlanRequest
) -> RusotoFuture<GetPlanResponse, GetPlanError>

Gets code to perform a specified mapping.

fn get_resource_policy(
&self
) -> RusotoFuture<GetResourcePolicyResponse, GetResourcePolicyError>

Retrieves a specified resource policy.

fn get_security_configuration(
&self,
input: GetSecurityConfigurationRequest
) -> RusotoFuture<GetSecurityConfigurationResponse, GetSecurityConfigurationError>

Retrieves a specified security configuration.

fn get_security_configurations(
&self,
input: GetSecurityConfigurationsRequest
) -> RusotoFuture<GetSecurityConfigurationsResponse, GetSecurityConfigurationsError>

Retrieves a list of all security configurations.

fn get_table(
&self,
input: GetTableRequest
) -> RusotoFuture<GetTableResponse, GetTableError>

Retrieves the Table definition in a Data Catalog for a specified table.

fn get_table_version(
&self,
input: GetTableVersionRequest
) -> RusotoFuture<GetTableVersionResponse, GetTableVersionError>

Retrieves a specified version of a table.

fn get_table_versions(
&self,
input: GetTableVersionsRequest
) -> RusotoFuture<GetTableVersionsResponse, GetTableVersionsError>

Retrieves a list of strings that identify available versions of a specified table.

fn get_tables(
&self,
input: GetTablesRequest
) -> RusotoFuture<GetTablesResponse, GetTablesError>

Retrieves the definitions of some or all of the tables in a given Database.

fn get_tags(
&self,
input: GetTagsRequest
) -> RusotoFuture<GetTagsResponse, GetTagsError>

Retrieves a list of tags associated with a resource.

fn get_trigger(
&self,
input: GetTriggerRequest
) -> RusotoFuture<GetTriggerResponse, GetTriggerError>

Retrieves the definition of a trigger.

fn get_triggers(
&self,
input: GetTriggersRequest
) -> RusotoFuture<GetTriggersResponse, GetTriggersError>

Gets all the triggers associated with a job.

fn get_user_defined_function(
&self,
input: GetUserDefinedFunctionRequest
) -> RusotoFuture<GetUserDefinedFunctionResponse, GetUserDefinedFunctionError>

Retrieves a specified function definition from the Data Catalog.

fn get_user_defined_functions(
&self,
input: GetUserDefinedFunctionsRequest
) -> RusotoFuture<GetUserDefinedFunctionsResponse, GetUserDefinedFunctionsError>

Retrieves a multiple function definitions from the Data Catalog.

fn import_catalog_to_glue(
&self,
input: ImportCatalogToGlueRequest
) -> RusotoFuture<ImportCatalogToGlueResponse, ImportCatalogToGlueError>

Imports an existing Athena Data Catalog to AWS Glue

fn list_crawlers(
&self,
input: ListCrawlersRequest
) -> RusotoFuture<ListCrawlersResponse, ListCrawlersError>

Retrieves the names of all crawler resources in this AWS account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

This operation takes the optional Tags field, which you can use as a filter on the response so that tagged resources can be retrieved as a group. If you choose to use tags filtering, only resources with the tag are retrieved.

fn list_dev_endpoints(
&self,
input: ListDevEndpointsRequest
) -> RusotoFuture<ListDevEndpointsResponse, ListDevEndpointsError>

Retrieves the names of all DevEndpoint resources in this AWS account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

This operation takes the optional Tags field, which you can use as a filter on the response so that tagged resources can be retrieved as a group. If you choose to use tags filtering, only resources with the tag are retrieved.

fn list_jobs(
&self,
input: ListJobsRequest
) -> RusotoFuture<ListJobsResponse, ListJobsError>

Retrieves the names of all job resources in this AWS account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

This operation takes the optional Tags field, which you can use as a filter on the response so that tagged resources can be retrieved as a group. If you choose to use tags filtering, only resources with the tag are retrieved.

fn list_triggers(
&self,
input: ListTriggersRequest
) -> RusotoFuture<ListTriggersResponse, ListTriggersError>

Retrieves the names of all trigger resources in this AWS account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

This operation takes the optional Tags field, which you can use as a filter on the response so that tagged resources can be retrieved as a group. If you choose to use tags filtering, only resources with the tag are retrieved.

fn put_data_catalog_encryption_settings(
&self,
input: PutDataCatalogEncryptionSettingsRequest
) -> RusotoFuture<PutDataCatalogEncryptionSettingsResponse, PutDataCatalogEncryptionSettingsError>

Sets the security configuration for a specified catalog. After the configuration has been set, the specified encryption is applied to every catalog write thereafter.

fn put_resource_policy(
&self,
input: PutResourcePolicyRequest
) -> RusotoFuture<PutResourcePolicyResponse, PutResourcePolicyError>

Sets the Data Catalog resource policy for access control.

fn reset_job_bookmark(
&self,
input: ResetJobBookmarkRequest
) -> RusotoFuture<ResetJobBookmarkResponse, ResetJobBookmarkError>

Resets a bookmark entry.

fn start_crawler(
&self,
input: StartCrawlerRequest
) -> RusotoFuture<StartCrawlerResponse, StartCrawlerError>

Starts a crawl using the specified crawler, regardless of what is scheduled. If the crawler is already running, returns a CrawlerRunningException.

fn start_crawler_schedule(
&self,
input: StartCrawlerScheduleRequest
) -> RusotoFuture<StartCrawlerScheduleResponse, StartCrawlerScheduleError>

Changes the schedule state of the specified crawler to SCHEDULED, unless the crawler is already running or the schedule state is already SCHEDULED.

fn start_job_run(
&self,
input: StartJobRunRequest
) -> RusotoFuture<StartJobRunResponse, StartJobRunError>

Starts a job run using a job definition.

fn start_trigger(
&self,
input: StartTriggerRequest
) -> RusotoFuture<StartTriggerResponse, StartTriggerError>

Starts an existing trigger. See Triggering Jobs for information about how different types of trigger are started.

fn stop_crawler(
&self,
input: StopCrawlerRequest
) -> RusotoFuture<StopCrawlerResponse, StopCrawlerError>

If the specified crawler is running, stops the crawl.

fn stop_crawler_schedule(
&self,
input: StopCrawlerScheduleRequest
) -> RusotoFuture<StopCrawlerScheduleResponse, StopCrawlerScheduleError>

Sets the schedule state of the specified crawler to NOT_SCHEDULED, but does not stop the crawler if it is already running.

fn stop_trigger(
&self,
input: StopTriggerRequest
) -> RusotoFuture<StopTriggerResponse, StopTriggerError>

Stops a specified trigger.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Adds tags to a resource. A tag is a label you can assign to an AWS resource. In AWS Glue, you can tag only certain resources. For information about what resources you can tag, see AWS Tags in AWS Glue.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Removes tags from a resource.

fn update_classifier(
&self,
input: UpdateClassifierRequest
) -> RusotoFuture<UpdateClassifierResponse, UpdateClassifierError>

Modifies an existing classifier (a GrokClassifier, an XMLClassifier, a JsonClassifier, or a CsvClassifier, depending on which field is present).

fn update_connection(
&self,
input: UpdateConnectionRequest
) -> RusotoFuture<UpdateConnectionResponse, UpdateConnectionError>

Updates a connection definition in the Data Catalog.

fn update_crawler(
&self,
input: UpdateCrawlerRequest
) -> RusotoFuture<UpdateCrawlerResponse, UpdateCrawlerError>

Updates a crawler. If a crawler is running, you must stop it using StopCrawler before updating it.

fn update_crawler_schedule(
&self,
input: UpdateCrawlerScheduleRequest
) -> RusotoFuture<UpdateCrawlerScheduleResponse, UpdateCrawlerScheduleError>

Updates the schedule of a crawler using a cron expression.

fn update_database(
&self,
input: UpdateDatabaseRequest
) -> RusotoFuture<UpdateDatabaseResponse, UpdateDatabaseError>

Updates an existing database definition in a Data Catalog.

fn update_dev_endpoint(
&self,
input: UpdateDevEndpointRequest
) -> RusotoFuture<UpdateDevEndpointResponse, UpdateDevEndpointError>

Updates a specified DevEndpoint.

fn update_job(
&self,
input: UpdateJobRequest
) -> RusotoFuture<UpdateJobResponse, UpdateJobError>

Updates an existing job definition.

fn update_partition(
&self,
input: UpdatePartitionRequest
) -> RusotoFuture<UpdatePartitionResponse, UpdatePartitionError>

Updates a partition.

fn update_table(
&self,
input: UpdateTableRequest
) -> RusotoFuture<UpdateTableResponse, UpdateTableError>

Updates a metadata table in the Data Catalog.

fn update_trigger(
&self,
input: UpdateTriggerRequest
) -> RusotoFuture<UpdateTriggerResponse, UpdateTriggerError>

Updates a trigger definition.

fn update_user_defined_function(
&self,
input: UpdateUserDefinedFunctionRequest
) -> RusotoFuture<UpdateUserDefinedFunctionResponse, UpdateUserDefinedFunctionError>

Updates an existing function definition in the Data Catalog.

Loading content...