Crate rusoto_glue

AWS Glue
Defines the public endpoint for the AWS Glue service.

If you're using the service, you're probably looking for GlueClient and Glue.

Structs

Action
Defines an action to be initiated by a trigger.

BatchCreatePartitionRequest

BatchCreatePartitionResponse

BatchDeleteConnectionRequest

BatchDeleteConnectionResponse

BatchDeletePartitionRequest

BatchDeletePartitionResponse

BatchDeleteTableRequest

BatchDeleteTableResponse

BatchDeleteTableVersionRequest

BatchDeleteTableVersionResponse

BatchGetPartitionRequest

BatchGetPartitionResponse

BatchStopJobRunError
Records an error that occurred when attempting to stop a specified job run.

BatchStopJobRunRequest

BatchStopJobRunResponse

BatchStopJobRunSuccessfulSubmission
Records a successful request to stop a specified JobRun.

CatalogEntry
Specifies a table definition in the Data Catalog.

CatalogImportStatus
A structure containing migration status information.

Classifier
Classifiers are triggered during a crawl task. A classifier checks whether a given file is in a format it can handle, and if it is, the classifier creates a schema in the form of a StructType object that matches that data format.

You can use the standard classifiers that AWS Glue supplies, or you can write your own classifiers to best categorize your data sources and specify the appropriate schemas to use for them. A classifier can be a grok classifier, an XML classifier, or a JSON classifier, as specified in one of the fields in the Classifier object.

CodeGenEdge
Represents a directional edge in a directed acyclic graph (DAG).

CodeGenNode
Represents a node in a directed acyclic graph (DAG)

CodeGenNodeArg
An argument or property of a node.

Column
A column in a Table.

Condition
Defines a condition under which a trigger fires.

Connection
Defines a connection to a data source.

ConnectionInput
A structure used to specify a connection to create or update.

ConnectionsList
Specifies the connections used by a job.

Crawler
Specifies a crawler program that examines a data source and uses classifiers to try to determine its schema. If successful, the crawler records metadata concerning the data source in the AWS Glue Data Catalog.

CrawlerMetrics
Metrics for a specified crawler.

CrawlerTargets
Specifies data stores to crawl.

CreateClassifierRequest

CreateClassifierResponse

CreateConnectionRequest

CreateConnectionResponse

CreateCrawlerRequest

CreateCrawlerResponse

CreateDatabaseRequest

CreateDatabaseResponse

CreateDevEndpointRequest

CreateDevEndpointResponse

CreateGrokClassifierRequest
Specifies a grok classifier for CreateClassifier to create.

CreateJobRequest

CreateJobResponse

CreateJsonClassifierRequest
Specifies a JSON classifier for CreateClassifier to create.

CreatePartitionRequest

CreatePartitionResponse

CreateScriptRequest

CreateScriptResponse

CreateTableRequest

CreateTableResponse

CreateTriggerRequest

CreateTriggerResponse

CreateUserDefinedFunctionRequest

CreateUserDefinedFunctionResponse

CreateXMLClassifierRequest
Specifies an XML classifier for CreateClassifier to create.

Database
The Database object represents a logical grouping of tables that may reside in a Hive metastore or an RDBMS.

DatabaseInput
The structure used to create or update a database.

DeleteClassifierRequest

DeleteClassifierResponse

DeleteConnectionRequest

DeleteConnectionResponse

DeleteCrawlerRequest

DeleteCrawlerResponse

DeleteDatabaseRequest

DeleteDatabaseResponse

DeleteDevEndpointRequest

DeleteDevEndpointResponse

DeleteJobRequest

DeleteJobResponse

DeletePartitionRequest

DeletePartitionResponse

DeleteTableRequest

DeleteTableResponse

DeleteTableVersionRequest

DeleteTableVersionResponse

DeleteTriggerRequest

DeleteTriggerResponse

DeleteUserDefinedFunctionRequest

DeleteUserDefinedFunctionResponse

DevEndpoint
A development endpoint where a developer can remotely debug ETL scripts.

DevEndpointCustomLibraries
Custom libraries to be loaded into a DevEndpoint.

DynamoDBTarget
Specifies a DynamoDB table to crawl.

ErrorDetail
Contains details about an error.

ExecutionProperty
An execution property of a job.

GetCatalogImportStatusRequest

GetCatalogImportStatusResponse

GetClassifierRequest

GetClassifierResponse

GetClassifiersRequest

GetClassifiersResponse

GetConnectionRequest

GetConnectionResponse

GetConnectionsFilter
Filters the connection definitions returned by the GetConnections API.

GetConnectionsRequest

GetConnectionsResponse

GetCrawlerMetricsRequest

GetCrawlerMetricsResponse

GetCrawlerRequest

GetCrawlerResponse

GetCrawlersRequest

GetCrawlersResponse

GetDatabaseRequest

GetDatabaseResponse

GetDatabasesRequest

GetDatabasesResponse

GetDataflowGraphRequest

GetDataflowGraphResponse

GetDevEndpointRequest

GetDevEndpointResponse

GetDevEndpointsRequest

GetDevEndpointsResponse

GetJobRequest

GetJobResponse

GetJobRunRequest

GetJobRunResponse

GetJobRunsRequest

GetJobRunsResponse

GetJobsRequest

GetJobsResponse

GetMappingRequest

GetMappingResponse

GetPartitionRequest

GetPartitionResponse

GetPartitionsRequest

GetPartitionsResponse

GetPlanRequest

GetPlanResponse

GetTableRequest

GetTableResponse

GetTableVersionRequest

GetTableVersionResponse

GetTableVersionsRequest

GetTableVersionsResponse

GetTablesRequest

GetTablesResponse

GetTriggerRequest

GetTriggerResponse

GetTriggersRequest

GetTriggersResponse

GetUserDefinedFunctionRequest

GetUserDefinedFunctionResponse

GetUserDefinedFunctionsRequest

GetUserDefinedFunctionsResponse

GlueClient
A client for the AWS Glue API.

GrokClassifier
A classifier that uses grok patterns.

ImportCatalogToGlueRequest

ImportCatalogToGlueResponse

JdbcTarget
Specifies a JDBC data store to crawl.

Job
Specifies a job definition.

JobBookmarkEntry
Defines a point which a job can resume processing.

JobCommand
Specifies code executed when a job is run.

JobRun
Contains information about a job run.

JobUpdate
Specifies information used to update an existing job definition. Note that the previous job definition will be completely overwritten by this information.

JsonClassifier
A classifier for JSON content.

LastCrawlInfo
Status and error information about the most recent crawl.

Location
The location of resources.

MappingEntry
Defines a mapping.

NotificationProperty
Specifies configuration properties of a notification.

Order
Specifies the sort order of a sorted column.

Partition
Represents a slice of table data.

PartitionError
Contains information about a partition error.

PartitionInput
The structure used to create and update a partion.

PartitionValueList
Contains a list of values defining partitions.

PhysicalConnectionRequirements
Specifies the physical requirements for a connection.

Predecessor
A job run that was used in the predicate of a conditional trigger that triggered this job run.

Predicate
Defines the predicate of the trigger, which determines when it fires.

ResetJobBookmarkRequest

ResetJobBookmarkResponse

ResourceUri
URIs for function resources.

S3Target
Specifies a data store in Amazon S3.

Schedule
A scheduling object using a cron statement to schedule an event.

SchemaChangePolicy
Crawler policy for update and deletion behavior.

Segment
Defines a non-overlapping region of a table's partitions, allowing multiple requests to be executed in parallel.

SerDeInfo
Information about a serialization/deserialization program (SerDe) which serves as an extractor and loader.

SkewedInfo
Specifies skewed values in a table. Skewed are ones that occur with very high frequency.

StartCrawlerRequest

StartCrawlerResponse

StartCrawlerScheduleRequest

StartCrawlerScheduleResponse

StartJobRunRequest

StartJobRunResponse

StartTriggerRequest

StartTriggerResponse

StopCrawlerRequest

StopCrawlerResponse

StopCrawlerScheduleRequest

StopCrawlerScheduleResponse

StopTriggerRequest

StopTriggerResponse

StorageDescriptor
Describes the physical storage of table data.

Table
Represents a collection of related data organized in columns and rows.

TableError
An error record for table operations.

TableInput
Structure used to create or update the table.

TableVersion
Specifies a version of a table.

TableVersionError
An error record for table-version operations.

Trigger
Information about a specific trigger.

TriggerUpdate
A structure used to provide information used to update a trigger. This object will update the the previous trigger definition by overwriting it completely.

UpdateClassifierRequest

UpdateClassifierResponse

UpdateConnectionRequest

UpdateConnectionResponse

UpdateCrawlerRequest

UpdateCrawlerResponse

UpdateCrawlerScheduleRequest

UpdateCrawlerScheduleResponse

UpdateDatabaseRequest

UpdateDatabaseResponse

UpdateDevEndpointRequest

UpdateDevEndpointResponse

UpdateGrokClassifierRequest
Specifies a grok classifier to update when passed to UpdateClassifier.

UpdateJobRequest

UpdateJobResponse

UpdateJsonClassifierRequest
Specifies a JSON classifier to be updated.

UpdatePartitionRequest

UpdatePartitionResponse

UpdateTableRequest

UpdateTableResponse

UpdateTriggerRequest

UpdateTriggerResponse

UpdateUserDefinedFunctionRequest

UpdateUserDefinedFunctionResponse

UpdateXMLClassifierRequest
Specifies an XML classifier to be updated.

UserDefinedFunction
Represents the equivalent of a Hive user-defined function (UDF) definition.

UserDefinedFunctionInput
A structure used to create or updata a user-defined function.

XMLClassifier
A classifier for XML content.

Enums

BatchCreatePartitionError
Errors returned by BatchCreatePartition

BatchDeleteConnectionError
Errors returned by BatchDeleteConnection

BatchDeletePartitionError
Errors returned by BatchDeletePartition

BatchDeleteTableError
Errors returned by BatchDeleteTable

BatchDeleteTableVersionError
Errors returned by BatchDeleteTableVersion

BatchGetPartitionError
Errors returned by BatchGetPartition

CreateClassifierError
Errors returned by CreateClassifier

CreateConnectionError
Errors returned by CreateConnection

CreateCrawlerError
Errors returned by CreateCrawler

CreateDatabaseError
Errors returned by CreateDatabase

CreateDevEndpointError
Errors returned by CreateDevEndpoint

CreateJobError
Errors returned by CreateJob

CreatePartitionError
Errors returned by CreatePartition

CreateScriptError
Errors returned by CreateScript

CreateTableError
Errors returned by CreateTable

CreateTriggerError
Errors returned by CreateTrigger

CreateUserDefinedFunctionError
Errors returned by CreateUserDefinedFunction

DeleteClassifierError
Errors returned by DeleteClassifier

DeleteConnectionError
Errors returned by DeleteConnection

DeleteCrawlerError
Errors returned by DeleteCrawler

DeleteDatabaseError
Errors returned by DeleteDatabase

DeleteDevEndpointError
Errors returned by DeleteDevEndpoint

DeleteJobError
Errors returned by DeleteJob

DeletePartitionError
Errors returned by DeletePartition

DeleteTableError
Errors returned by DeleteTable

DeleteTableVersionError
Errors returned by DeleteTableVersion

DeleteTriggerError
Errors returned by DeleteTrigger

DeleteUserDefinedFunctionError
Errors returned by DeleteUserDefinedFunction

GetCatalogImportStatusError
Errors returned by GetCatalogImportStatus

GetClassifierError
Errors returned by GetClassifier

GetClassifiersError
Errors returned by GetClassifiers

GetConnectionError
Errors returned by GetConnection

GetConnectionsError
Errors returned by GetConnections

GetCrawlerError
Errors returned by GetCrawler

GetCrawlerMetricsError
Errors returned by GetCrawlerMetrics

GetCrawlersError
Errors returned by GetCrawlers

GetDatabaseError
Errors returned by GetDatabase

GetDatabasesError
Errors returned by GetDatabases

GetDataflowGraphError
Errors returned by GetDataflowGraph

GetDevEndpointError
Errors returned by GetDevEndpoint

GetDevEndpointsError
Errors returned by GetDevEndpoints

GetJobError
Errors returned by GetJob

GetJobRunError
Errors returned by GetJobRun

GetJobRunsError
Errors returned by GetJobRuns

GetJobsError
Errors returned by GetJobs

GetMappingError
Errors returned by GetMapping

GetPartitionError
Errors returned by GetPartition

GetPartitionsError
Errors returned by GetPartitions

GetPlanError
Errors returned by GetPlan

GetTableError
Errors returned by GetTable

GetTableVersionError
Errors returned by GetTableVersion

GetTableVersionsError
Errors returned by GetTableVersions

GetTablesError
Errors returned by GetTables

GetTriggerError
Errors returned by GetTrigger

GetTriggersError
Errors returned by GetTriggers

GetUserDefinedFunctionError
Errors returned by GetUserDefinedFunction

GetUserDefinedFunctionsError
Errors returned by GetUserDefinedFunctions

GlueBatchStopJobRunError
Errors returned by BatchStopJobRun

ImportCatalogToGlueError
Errors returned by ImportCatalogToGlue

ResetJobBookmarkError
Errors returned by ResetJobBookmark

StartCrawlerError
Errors returned by StartCrawler

StartCrawlerScheduleError
Errors returned by StartCrawlerSchedule

StartJobRunError
Errors returned by StartJobRun

StartTriggerError
Errors returned by StartTrigger

StopCrawlerError
Errors returned by StopCrawler

StopCrawlerScheduleError
Errors returned by StopCrawlerSchedule

StopTriggerError
Errors returned by StopTrigger

UpdateClassifierError
Errors returned by UpdateClassifier

UpdateConnectionError
Errors returned by UpdateConnection

UpdateCrawlerError
Errors returned by UpdateCrawler

UpdateCrawlerScheduleError
Errors returned by UpdateCrawlerSchedule

UpdateDatabaseError
Errors returned by UpdateDatabase

UpdateDevEndpointError
Errors returned by UpdateDevEndpoint

UpdateJobError
Errors returned by UpdateJob

UpdatePartitionError
Errors returned by UpdatePartition

UpdateTableError
Errors returned by UpdateTable

UpdateTriggerError
Errors returned by UpdateTrigger

UpdateUserDefinedFunctionError
Errors returned by UpdateUserDefinedFunction

Traits

Glue
Trait representing the capabilities of the AWS Glue API. AWS Glue clients implement this trait.