Struct rusoto_glue::Crawler

pub struct Crawler {
    pub classifiers: Option<Vec<String>>,
    pub configuration: Option<String>,
    pub crawl_elapsed_time: Option<i64>,
    pub creation_time: Option<f64>,
    pub database_name: Option<String>,
    pub description: Option<String>,
    pub last_crawl: Option<LastCrawlInfo>,
    pub last_updated: Option<f64>,
    pub name: Option<String>,
    pub role: Option<String>,
    pub schedule: Option<Schedule>,
    pub schema_change_policy: Option<SchemaChangePolicy>,
    pub state: Option<String>,
    pub table_prefix: Option<String>,
    pub targets: Option<CrawlerTargets>,
    pub version: Option<i64>,
}

Specifies a crawler program that examines a data source and uses classifiers to try to determine its schema. If successful, the crawler records metadata concerning the data source in the AWS Glue Data Catalog.

Fields
classifiers: Option<Vec<String>>

A list of custom classifiers associated with the crawler.

configuration: Option<String>

Crawler configuration information. This versioned JSON string allows users to specify aspects of a Crawler's behavior.

You can use this field to force partitions to inherit metadata such as classification, input format, output format, serde information, and schema from their parent table, rather than detect this information separately for each partition. Use the following JSON string to specify that behavior:

Example: '{ "Version": 1.0, "CrawlerOutput": { "Partitions": { "AddOrUpdateBehavior": "InheritFromTable" } } }'

crawl_elapsed_time: Option<i64>

If the crawler is running, contains the total time elapsed since the last crawl began.

creation_time: Option<f64>

The time when the crawler was created.

database_name: Option<String>

The database where metadata is written by this crawler.

description: Option<String>

A description of the crawler.

last_crawl: Option<LastCrawlInfo>

The status of the last crawl, and potentially error information if an error occurred.

last_updated: Option<f64>

The time the crawler was last updated.

name: Option<String>

The crawler name.

role: Option<String>

The IAM role (or ARN of an IAM role) used to access customer resources, such as data in Amazon S3.

schedule: Option<Schedule>

For scheduled crawlers, the schedule when the crawler runs.

schema_change_policy: Option<SchemaChangePolicy>

Sets the behavior when the crawler finds a changed or deleted object.

state: Option<String>

Indicates whether the crawler is running, or whether a run is pending.

table_prefix: Option<String>

The prefix added to the names of tables that are created.

targets: Option<CrawlerTargets>

A collection of targets to crawl.

version: Option<i64>

The version of the crawler.